# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Timur Karimov Resource Library API",
    "version": "1.0.0",
    "description": "API documentation created by Timur Karimov"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local server"
    }
  ],
  "paths": {
    "/api/v1/events": {
      "get": {
        "summary": "Get all events",
        "description": "Returns a list of all events in the system.",
        "tags": [
          "Events"
        ],
        "responses": {
          "200": {
            "description": "Events retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new event",
        "description": "Creates a new event using the provided request body.",
        "tags": [
          "Events"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSingleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/events/{id}": {
      "get": {
        "summary": "Get an event by ID",
        "description": "Returns one event that matches the provided ID.",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The unique ID of the event",
            "schema": {
              "type": "string",
              "example": "aB3dEfGhIjKlMnOpQrSt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSingleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Event not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update an event",
        "description": "Updates an existing event that matches the provided ID.",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The unique ID of the event",
            "schema": {
              "type": "string",
              "example": "aB3dEfGhIjKlMnOpQrSt"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventSingleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Event not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete an event",
        "description": "Deletes an existing event that matches the provided ID.",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The unique ID of the event",
            "schema": {
              "type": "string",
              "example": "aB3dEfGhIjKlMnOpQrSt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDeleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Event not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/health": {
      "get": {
        "summary": "Check API health status",
        "description": "Returns the current health status of the API.",
        "tags": [
          "Health"
        ],
        "responses": {
          "200": {
            "description": "API is running successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "uptime": {
                      "type": "number",
                      "example": 123.456
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2026-03-28T18:30:00.000Z"
                    },
                    "version": {
                      "type": "string",
                      "example": "v1"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Event": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "aB3dEfGhIjKlMnOpQrSt"
          },
          "name": {
            "type": "string",
            "example": "Spring Tech Meetup"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-15T18:00:00.000Z"
          },
          "capacity": {
            "type": "integer",
            "example": 100
          },
          "registrationCount": {
            "type": "integer",
            "example": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "cancelled",
              "completed"
            ],
            "example": "active"
          },
          "category": {
            "type": "string",
            "enum": [
              "conference",
              "workshop",
              "meetup",
              "seminar",
              "general"
            ],
            "example": "meetup"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateEventRequest": {
        "type": "object",
        "required": [
          "name",
          "date",
          "capacity"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Spring Tech Meetup"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-15T18:00:00.000Z"
          },
          "capacity": {
            "type": "integer",
            "example": 100
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "cancelled",
              "completed"
            ],
            "example": "active"
          },
          "category": {
            "type": "string",
            "enum": [
              "conference",
              "workshop",
              "meetup",
              "seminar",
              "general"
            ],
            "example": "meetup"
          }
        }
      },
      "UpdateEventRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Spring Tech Meetup Updated"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-16T18:00:00.000Z"
          },
          "capacity": {
            "type": "integer",
            "example": 150
          },
          "registrationCount": {
            "type": "integer",
            "example": 10
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "cancelled",
              "completed"
            ],
            "example": "active"
          },
          "category": {
            "type": "string",
            "enum": [
              "conference",
              "workshop",
              "meetup",
              "seminar",
              "general"
            ],
            "example": "workshop"
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "\"name\" is required",
              "\"date\" must be in ISO 8601 date format"
            ]
          }
        }
      },
      "EventListResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Events retrieved successfully"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            }
          }
        }
      },
      "EventSingleResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Event retrieved successfully"
          },
          "data": {
            "$ref": "#/components/schemas/Event"
          }
        }
      },
      "EventDeleteResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Event deleted successfully"
          }
        }
      }
    }
  },
  "tags": []
}